+++ persistence/sqlpersistence/sqlite/event.go
@@ -206,7 +206,9 @@ func (driver) SelectNextEventOffset(
 		ctx,
 		`SELECT
 			next_offset
-		FROM event_offset`,
+		FROM event_offset
+		WHERE source_app_key = $1`,
+		ak,
 	)
 
 	var next uint64